Guide Addition на DTO+

+++ app/Modules/Guide/Application/Actions/Addition/RemoveAdditionUseCase.php
@@ -3,8 +3,10 @@
 namespace App\Modules\Guide\Application\Actions\Addition;
 
 use App\Modules\Guide\Domain\Interfaces\AdditionRepositoryInterface;
+use App\Modules\Shared\Domain\Entities\UserPermission;
+use App\Modules\Shared\Domain\Exceptions\AccessDeniedException;
 
-class RemoveAdditionUseCase
+readonly class RemoveAdditionUseCase
 {
 
     public function __construct(private AdditionRepositoryInterface $repository)
@@ -12,9 +14,9 @@ class RemoveAdditionUseCase
 
     }
 
-    public function execute(int $id)
+    public function execute(int $id, UserPermission $permission): void
     {
-
+        if (!$permission->can('guide.guide.remove')) throw new AccessDeniedException();
         //TODO Проверка на наличие услуг
 
         $this->repository->remove($id);
